+++ src/functionalTest/java/uk/gov/hmcts/reform/em/hrs/functional/util/ExtendedCcdHelper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.em.hrs.testutil;
+package uk.gov.hmcts.reform.em.hrs.functional.util;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -11,10 +11,10 @@ import uk.gov.hmcts.reform.em.test.ccddefinition.CcdDefImportApi;
 import uk.gov.hmcts.reform.em.test.ccddefinition.CcdDefUserRoleApi;
 import uk.gov.hmcts.reform.em.test.idam.IdamHelper;
 
-import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
+import javax.annotation.PostConstruct;
 
 @Service
 public class ExtendedCcdHelper {
